backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis[hiredis]==5.2.0
httpx==0.28.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.10.0
pydantic-settings==2.6.0
python-multipart==0.0.12
websockets==14.1
langchain==0.3.0
langchain-openai==0.2.0
python-dotenv==1.0.1
tenacity==9.0.0
slowapi==0.1.9
